+++ NEWS.md
@@ -70,7 +70,8 @@
 * Added `"viridisE"` (the [cividis](https://github.com/marcosci/cividis) 
   version of viridis) to the supported colors. 
 
-* `ppc_bars()` and `ppc_bars_grouped()` now allow negative integers as input. (#172)
+* `ppc_bars()` and `ppc_bars_grouped()` now allow negative integers as input.
+  (#172, @jeffpollock9)
 
 
 # bayesplot 1.6.0